requirements.txt
asgiref==3.4.1
Django==3.2.15
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
psycopg2-binary==2.9.3
PyJWT==1.7.1
python-decouple==3.6
pytz==2022.4
sqlparse==0.4.3
typing_extensions==4.1.1
boto3==1.23.10
xlrd==1.2.0